+++ scripts/Build-Installer.ps1
@@ -1,6 +1,6 @@
 ﻿[CmdletBinding()]
 param(
-    [string]$Version = "1.2.1",
+    [string]$Version = "1.2.3",
     [string]$WixCommand = "",
     [string]$ArtifactsDirectory = ""
 )